inference-service/requirements.txt
Flask
flask_cors
numpy
faiss-cpu
h5py
psycopg2
python-dotenv
flask_apscheduler
ijson
flasgger
opencv-python
requests
tqdm
filelock
watchdog
flask-limiter[redis]
huggingface_hub
redis